"Commit" to save the data
