+++ app/build.gradle
@@ -48,6 +48,9 @@ dependencies {
     //add twitter dependency
     //implementation 'com.twitter.sdk.android:twitter-core:3.1.0'
 
+    //add Motion layout dependency
+    implementation 'androidx.constraintlayout:constraintlayout:2.0.0-beta8'
+
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'androidx.appcompat:appcompat:1.1.0'
@@ -55,6 +58,7 @@ dependencies {
     implementation 'androidx.navigation:navigation-fragment:2.3.0'
     implementation 'androidx.navigation:navigation-ui:2.3.0'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
+    implementation 'androidx.preference:preference:1.1.1'
     testImplementation 'junit:junit:4.13'
     androidTestImplementation 'androidx.test.ext:junit:1.1.1'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
