+++ app/src/main/java/com/example/final_project/Daatabase/useradate/BlankContract.java
@@ -1,4 +1,4 @@
-package com.example.final_project.Daatabase;
+package com.example.final_project.Daatabase.useradate;
 
 import android.provider.BaseColumns;
