+++ app/src/main/java/com/example/final_project/Daatabase/useradate/UserDatadbProvider.java
@@ -1,4 +1,4 @@
-package com.example.final_project.Daatabase;
+package com.example.final_project.Daatabase.useradate;
 
 import android.database.Cursor;
 import android.database.SQLException;
