+++ app/src/main/res/layout/fragment_payments.xml
@@ -2,12 +2,13 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.final_project.PaymentsFragment">
+    tools:context="com.example.final_project.ui.payments.PaymentsFragment">
 
     <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <!--<TextView
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:text="payments" />
+        android:text="payments" />-->
+
 
 </FrameLayout>
