+++ app/src/main/res/values/strings.xml
@@ -1,22 +1,31 @@
 <resources>
     <string name="app_name">Final_Project</string>
     <string name="title_activity_main">MainActivity</string>
-    <string name="navigation_drawer_open">Open navigation drawer</string>
-    <string name="navigation_drawer_close">Close navigation drawer</string>
-    <string name="nav_header_title">Android Studio</string>
-    <string name="nav_header_subtitle">android.studio@android.com</string>
-    <string name="nav_header_desc">Navigation header</string>
     <string name="action_settings">Settings</string>
     <string name="action_sign_out">Sign Out</string>
-
-    <string name="menu_home">Home</string>
-    <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
 
     <string name="twitter_consumer_key" translatable="false">uuOjlXdSNioHKcJjldMlE1kvo</string>
     <string name="twitter_consumer_secret" translatable="false">w6qRy3wiAkngl51ghggl8MeEyBMPLM0Zd2fnFZuTVifjGIg5iH</string>
     <string name="twitter_bearer_token" translatable="false">wAAAAAAAAAAAAAAAAAAAAANskGAEAAAAAFH6EH9Oyv6X4%2Bh9gBIVt8ecX6%2Bo%3DbGHIpvaEmxyhouJbArW2tS8EO7VNMGvzl424KriUN3dO23Yum6</string>
+
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
+    <string name="title_due_pament">Due Payment</string>
+    <string name="title_setting">Setting</string>
+    <string name="title_home">Home</string>
+    <!-- Preference Titles -->
+    <string name="messages_header">Messages</string>
+    <string name="sync_header">Sync</string>
+
+    <!-- Messages Preferences -->
+    <string name="signature_title">Your signature</string>
+    <string name="reply_title">Default reply action</string>
+
+    <!-- Sync Preferences -->
+    <string name="sync_title">Sync email periodically</string>
+    <string name="attachment_title">Download incoming attachments</string>
+    <string name="attachment_summary_on">Automatically download attachments for incoming emails
+    </string>
+    <string name="attachment_summary_off">Only download attachments when manually requested</string>
 
 </resources>
